+++ ISHPermissionKit/Requests/ISHPermissionRequest.m
@@ -7,7 +7,9 @@
 //
 
 #import "ISHPermissionRequest.h"
+#import "ISHPermissionRequest+Private.h"
 @import CoreLocation;
+@import UIKit;
 
 @interface ISHPermissionRequest ()
 @property (nonatomic, readwrite) ISHPermissionCategory permissionCategory;
@@ -25,6 +27,10 @@
 }
 
 - (BOOL)mayRequestUserPermissionWithCompletionBlock:(ISHPermissionRequestCompletionBlock)completion {
+#if DEBUG
+    [self verifyUsageDescriptions];
+#endif
+
     if (!completion) {
         NSAssert(completion, @"requestUserPermissionWithCompletionBlock: requires a completion block");
         return NO;
@@ -70,6 +76,27 @@
     }
     return error;
 }
+
+#pragma mark - Usage Key Verification
+
+#if DEBUG
+- (NSArray<NSString *> *)staticUsageDescriptionKeys {
+    return nil;
+}
+
+- (void)verifyUsageDescriptions {
+    // we need to get the app's info plist, not the framework's
+    Class delegateClass = [UIApplication sharedApplication].delegate.class;
+    NSBundle *appBundle = [NSBundle bundleForClass:delegateClass];
+    NSDictionary *infoPlist = appBundle.infoDictionary;
+
+    for (NSString *usageKey in [self staticUsageDescriptionKeys]) {
+        id content = infoPlist[usageKey];
+        NSString *stringContent = [content isKindOfClass:[NSString class]] ? (NSString *)content : nil;
+        NSAssert(stringContent.length > 0, @"You must provide %@ in your info plist when using %@", usageKey, ISHStringFromPermissionCategory(self.permissionCategory));
+    }
+}
+#endif
 
 @end