+++ ISHPermissionKit/Requests/ISHPermissionRequestPhotoLibrary.m
@@ -51,4 +51,10 @@
     }];
 }
 
+#if DEBUG
+- (NSArray<NSString *> *)staticUsageDescriptionKeyss {
+    return @[@"NSPhotoLibraryUsageDescription"];
+}
+#endif
+
 @end